+++ CHANGES.md
@@ -4,10 +4,15 @@
 - Replaced the `@ReadOnly` annotation with a `@Field(FieldType.PROTECTED)` (see [#108](https://github.com/klum-dsl/klum-ast/issues/108))
 - Included a TRANSIENT field type to allow fields that are not part of the model (see [#102](https://github.com/klum-dsl/klum-ast/issues/102))
 - PROTECTED (former ReadOnly) fields now DO create adder methods, although protected (see [#78](https://github.com/klum-dsl/klum-ast/issues/78))
+- Issue a warning if a field is named `owner`
 - __(Potentially) breaking changes:__
     - Introduce a new createFromClasspath method (see [#110](https://github.com/klum-dsl/klum-ast/issues/110))
     - this means that either klum-ast or future klum-util package needs to be present in the classpath during
       runtime
+    - Closures are all `DelegateOnly` instead of the previous `DelegateFirst`. This means that you cannot access
+      methods of an outer object directly (which would not be very intuitive). If you need this functionality,
+      you need to access the outer object directly, using the `owner` property of `Closure` or an `@Owner` field
+      of the outer instance. See [[Migration]] for details. ([#72](https://github.com/klum-dsl/klum-ast/issues/72))
 
 
 ## 1.1.1